+++ packages/orq-rc/src/lib/config.ts
@@ -64,7 +64,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "3.2.0-rc.12",
+  sdkVersion: "3.2.0-rc.13",
   genVersion: "2.524.1",
-  userAgent: "speakeasy-sdk/typescript 3.2.0-rc.12 2.524.1 2.0 @orq-ai/node",
+  userAgent: "speakeasy-sdk/typescript 3.2.0-rc.13 2.524.1 2.0 @orq-ai/node",
 } as const;