+++ packages/orq-rc/RELEASES.md
@@ -539,3 +539,13 @@ Based on:
 - [typescript v3.2.0-rc.12] packages/orq-rc
 ### Releases
 - [NPM v3.2.0-rc.12] https://www.npmjs.com/package/@orq-ai/node/v/3.2.0-rc.12 - packages/orq-rc
+
+## 2025-02-25 11:31:04
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.502.0 (2.524.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.2.0-rc.13] packages/orq-rc
+### Releases
+- [NPM v3.2.0-rc.13] https://www.npmjs.com/package/@orq-ai/node/v/3.2.0-rc.13 - packages/orq-rc